Multiple pose bug fix & output visualizations in the image demo (#29)
This PR fixes a bug in multi-pose estimation that caused inaccuracies and jitteriness.  
It also adds visualizations of the model outputs to the images demo.

I had been working on a modification to the "coco" demo that allows the outputs used for multiple-pose estimation to be visualized.  This would show the heatmap (in yellow), short-range offsets (in red) and mid-range offset vectors (in blue) for a specific part, with the opacity is determined by the score of the corresponding heatmap point.  Note that the midrange vector origin is the heatmap coordinate + short-range offset.  Here it is on the tennis example, with the left knee selected and an **output stride of 8**:

![visualization](https://user-images.githubusercontent.com/891755/41256091-9bae4082-6d96-11e8-8473-8bc156d7be2a.png)

These new visualizations above are included as part of the PR.

For details of the multi-pose estimation algorithm, **refer to the research paper [PersonLab: Person Pose Estimation and Instance Segmentation with a Bottom-Up, Part-Based, Geometric Embedding Model](https://arxiv.org/pdf/1803.08225)** and the section `Fast greedy decoding` on which it is based, or the PoseNet [decodeMultiplePoses function](https://github.com/tensorflow/tfjs-models/blob/master/posenet/src/multiPose/decodeMultiplePoses.ts#L117) where it is implemented.

When **turning the output stride up to 32**, the innacuracy issues for multi-pose estimation became clear.   Below, the left image shows the mid-range offsets for the right shoulder, and the right image shows the heatmap and short-range offsets for the right elbow for the backpack image:

![badoffsets](https://user-images.githubusercontent.com/891755/41256286-2357eca4-6d97-11e8-8f63-593fd428a257.png)

As can be seen above, even though the displacement vectors for the right shoulder are pointing near the right elbow, and the short-range offset vectors are pointing at the right elbow, the estimated keypoint for the right elbow is way off. 

Essentially what the code was doing is:

```javascript
point = lastKeypoint + midRangeOffset
nearestHeatmapIndex = getNearestHeatmapIndexAtPoint(point)
offsetVector = getOffsetVectorAtIndex(nearestHeatmapIndex)

estimatedKeypoint =  lastKeypoint + midRangeOffset + offsetVector
```

What this means is essentially the short range's offset was considered to have an origin at the end of the last displacement vector, instead of from the heatmap point.  When adding the vectors this way, the final estimated keypoint is off from where it should be.

After changing this to have the offset originate from the heatmap point:
```javascript
estimatedKeypoint =  nearestHeatmapIndex * outputStride + offsetVector
```

The results are improved significantly:

![fixed](https://user-images.githubusercontent.com/891755/41256710-690f4944-6d98-11e8-8704-c94cd3fed2a9.png)

This fix reduces much of the jitteriness and inaccuracies that occured in multi-pose estimation, especially when using the camera.


This PR also:
* For the camera demo, changes it to default to multi-pose since that is now stable, and loads the 0.75 model by default since that gets reasonable real-time performance on my Macbook.
* Ignores `yalc` files in the case someone is developing locally and wants to use yalc to link dependencies to local posenet.
* exports the `poseChain` from multi-pose estimation so that it can be used in the image demo.
* updates the tfjs version of the camera demo to get around a bug [that occured in fromPixels](https://github.com/tensorflow/tfjs/issues/349)