Person and Body Part Segmentation model (#63)
This PR adds support for loading and decoding new pre-trained models provided by Google researcher Tyler Zhu (@tylerzhu-github) which does person segmentation and body part mapping on an image.  It also adds demos showing how they can be visualized.

The new demos can be tried here:
https://person-segmentation-demos.netlify.com

Here is segmentation and estimated part map with output stride of 8:  The dark gray area indicates what is outside of the segmentation area:
![download 3](https://user-images.githubusercontent.com/891755/46114169-77587400-c1bf-11e8-90cb-4f4ea1e504e9.png)
![download 4](https://user-images.githubusercontent.com/891755/46114245-c30b1d80-c1bf-11e8-9452-96877a54ba7d.png)

Here is purely a segmentation output with output stride of 16.  The dark gray area indicates what is outside of the segmentation area:
![download 2](https://user-images.githubusercontent.com/891755/46114118-3fe9c780-c1bf-11e8-81d3-138d68f2ff92.png)


## The Model

This new model provides two outputs: a 2d tensor for segmentation, and a 3d tensor used for body part maps.  It requires an image size of 257x353.   It only supports segmentation on a single person. However, if combined with a person box detector in a first stage, it can be used in a second stage to do segmentation and part maps for each bounding box, therefore becoming a multi-person segment and part mapper estimator.  

Unlike Posenet's pose estimation, which has 17 keypoints, the part map refers to 24 *part channels*:

```typescript
export const partChannels: string[] = [
  'left_face',
  'right_face',
  'right_upper_leg_front',
  'right_lower_leg_back',
  'right_upper_leg_back',
  'left_lower_leg_front',
  'left_upper_leg_front',
  'left_upper_leg_back',
  'left_lower_leg_back',
  'right_feet',
  'right_lower_leg_front',
  'left_feet',
  'torso_front',
  'torso_back',
  'right_upper_arm_front',
  'right_upper_arm_back',
  'right_lower_arm_back',
  'left_lower_arm_front',
  'left_upper_arm_front',
  'left_upper_arm_back',
  'left_lower_arm_back',
  'right_hand',
  'right_lower_arm_front',
  'left_hand'
];
```

The 3d tensor output for body part map's 3rd axis corresponds to each of these part channels, and for each part channel has a 2d tensor which indicates the probability that a part channel exists in each area.

There are 4 MobileNet-architecture models available as part of this PR, with 1.00, 0.75, 0.50, and 0.25 depth multipliers.  The higher the depth multiplier the higher the accuracy and slower the speed, and visa versa with lower depth multipliers.

These new models are in the frozen graph format, so the PoseNet code was modified to support loading weights either from a frozen graph format or the legacy checkpoint format which is what the existing posenet models use.

The model is loaded like:

```typescript
export async function load(multiplier: MobileNetMultiplier = 0.75)
```

Unlike the existing PoseNet `load` function, this loads using the frozen graph format. 

## The new API

This new model, called PersonSegmentation, provides three api methods:

```typescript
async estimatePersonSegmentation(
      input: InputType, flipHorizontal = false, outputStride: OutputStride = 16,
      segmentationThreshold = 0.5): Promise<Int32Array>

async estimatePartSegmentation(
      input: InputType, flipHorizontal = false, outputStride: OutputStride = 16,
      segmentationThreshold = 0.5,
      partColors: Array<[number, number, number]>): Promise<Int32Array>
```

The segmentation mask returned from `estimatePersonSegmentation` is an array with values 0 or 1 corresponding to if a person was estimated to be in each pixel. The array size corresponds to the number of pixels in the image. See the readme for the body part ids.  If a value is 1, a person was estimated to be in the corresponding pixel.

The returned value from estimatePartSegmentation is an array with values 0-24 corresponding to the body part for each pixel, indexed by pixel h, w.  The array size corresponds to the number of pixels in the image. See the readme for the body part ids.  If a value is -1, no body part was estimated to be in that pixel.

The model handles resizing and padding the image to match the **required resolution of 257x353** while maintaining the aspect ratio before feeding it through the network.  It also scales the outputs back up to the original input resolution while cropping out the padding before returning the outputs.  

## The demos

This includes new demos for segmentation and part maps that only work with a single person.

The demos can be viewed here:
https://person-segmentation-demos.netlify.com